requirements.txt
dhanhq==2.0.0
websocket-client==1.6.4
python-dotenv==1.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
requests>=2.31.0
websocket-server==0.4
flask>=2.0.0
flask-cors>=3.0.0 